fix: Resolve CI failures for source-gen JSON PR
- Fix Directory.Build.props: use $(MSBuildThisFileDirectory) instead of
  $(SolutionDir) so Global/Usings.cs resolves in project-level builds
  (fixes Benchmark and E2E CI failures)
- Extract DispatchData/DispatchSubscriptionData into Runtime.Dispatch.cs
  partial class to isolate source-gen JSON changes
- Auto-format Runtime.cs to pass dotnet format lint check (fixes
  pre-existing whitespace, unused usings, name simplifications)
